Modificaciones de isnstructivos

+++ app/Http/Controllers/InstructivoController.php
@@ -2,12 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use DB;
+Use Session;
+use App\Instructivo;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
-use App\Instructivo;
-Use Session;
-use DB;
-
 
 class InstructivoController extends Controller{
 
@@ -31,26 +30,36 @@ class InstructivoController extends Controller{
         return view('instructivos.create');       
     }
 
-    public function store_instructivo(Request $request){        
+    public function store_instructivo(Request $request)
+    {
+        $maxFileSize = 41943040; // Tamaño máximo permitido en bytes (40 MB en este caso)
         $aux = Instructivo::get()->max('id');
-        if($aux == null){
+        if ($aux == null) {
             $aux = 0;
         }
-
+        
         $instructivo = new Instructivo;
         $instructivo->titulo = $request['tituloCreate'];
         $instructivo->tipo = $request['tipo_instructivo'];
-
-        if($request->file('archivo')){
+        
+        if ($request->hasFile('archivo')) {
             $file = $request->file('archivo');
-            $name = str_pad($aux + 1, 5, '0', STR_PAD_LEFT).$file->getClientOriginalName();         
-            Storage::disk('public')->put('instructivo/'.$name, \File::get($file));
-            $instructivo->archivo = 'instructivo\\'.$name;
-        }
-
+            
+            if ($file->getSize() > $maxFileSize) {
+                // El archivo excede el tamaño máximo permitido
+                Session::flash('message', 'El archivo excede el tamaño máximo permitido de 40 MB');
+                Session::flash('alert-class', 'alert-danger');
+                return redirect('instructivos');
+            }
+            
+            $name = str_pad($aux + 1, 5, '0', STR_PAD_LEFT) . $file->getClientOriginalName();
+            Storage::disk('public')->put('instructivo/' . $name, \File::get($file));
+            $instructivo->archivo = 'instructivo\\' . $name;
+        }    
+        
         $instructivo->save();
-
-        Session::flash('message','Archivo agregado con éxito');
+        
+        Session::flash('message', 'Archivo agregado con éxito');
         Session::flash('alert-class', 'alert-success');
         return redirect('instructivos');
     }
@@ -72,27 +81,42 @@ class InstructivoController extends Controller{
         return view('instructivos.update', ['instructivo' => $instructivo]);
     }
 
-    public function update_instructivo(Request $request){
-        if($request['tituloUpdate'] or $request['tipo_instructivo']){
+    public function update_instructivo(Request $request)
+    {
+        if ($request['tituloUpdate'] || $request['tipo_instructivo']) {
             $instructivo = DB::table('instructivo')
-            ->where('instructivo.id',$request['id'])
-            ->update(['titulo' => $request['tituloUpdate'], 'tipo' => $request['tipo_instructivo']]);        
+                ->where('instructivo.id', $request['id'])
+                ->update(['titulo' => $request['tituloUpdate'], 'tipo' => $request['tipo_instructivo']]);
         }
+        
         $aux = Instructivo::find($request['id']);
-        if($request['archivo'] != null){
-            if($request->file('archivo')){
-                if ($aux->archivo != null){
-                    unlink(storage_path('app\\public\\'.$aux->archivo));
-                }
+        
+        if ($request['archivo'] != null) {
+            if ($request->file('archivo')) {
+                $maxFileSize = 41943040; // Tamaño máximo permitido en bytes (40 MB en este caso)
                 $file = $request->file('archivo');
-                $name = str_pad($request['id'], 5, '0', STR_PAD_LEFT).$file->getClientOriginalName();
-                Storage::disk('public')->put('instructivo/'.$name, \File::get($file));
+                
+                if ($file->getSize() > $maxFileSize) {
+                    // El archivo excede el tamaño máximo permitido
+                    Session::flash('message', 'El archivo excede el tamaño máximo permitido de 40 MB');
+                    Session::flash('alert-class', 'alert-danger');
+                    return redirect('instructivos');
+                }
+                
+                if ($aux->archivo != null) {
+                    unlink(storage_path('app\\public\\' . $aux->archivo));
+                }
+                
+                $name = str_pad($request['id'], 5, '0', STR_PAD_LEFT) . $file->getClientOriginalName();
+                Storage::disk('public')->put('instructivo/' . $name, \File::get($file));
+                
                 $instructivo = DB::table('instructivo')
-                ->where('instructivo.id',$request['id'])
-                ->update(['archivo' => 'instructivo\\'.$name]);
+                    ->where('instructivo.id', $request['id'])
+                    ->update(['archivo' => 'instructivo\\' . $name]);
             }
         }
-        Session::flash('message','Archivo modificado con éxito');
+        
+        Session::flash('message', 'Archivo modificado con éxito');
         Session::flash('alert-class', 'alert-success');
         return redirect('instructivos');
     }
